index -> app refactor
