+++ auth/src/index.ts
@@ -1,48 +1,7 @@
-import express, { json } from 'express';
-import 'express-async-errors';
 import mongoose from 'mongoose';
-import cookieSession from 'cookie-session';
-
-import { errorHandler } from './middleware';
-import {
-  signinRouter,
-  signoutRouter,
-  signupRouter,
-  currentUserRouter,
-} from './routes';
-import { NotFoundError } from './errors';
+import app from './app';
 
 const port = process.env.PORT || 3000;
-
-/**
- * App setup
- */
-const app = express();
-app.set('trust proxy', true);
-app.use(json());
-app.use(
-  cookieSession({
-    signed: false,
-    secure: true,
-  })
-);
-
-/**
- * Routes
- */
-app.use(currentUserRouter);
-app.use(signinRouter);
-app.use(signoutRouter);
-app.use(signupRouter);
-
-app.all(`*`, async (req, res) => {
-  throw new NotFoundError();
-});
-
-/**
- * Error Handler
- */
-app.use(errorHandler);
 
 const start = async () => {
   if (!process.env.JWT_KEY) {
